+++ univention-ox-soap-api/modules/univention/ox/soap/config.py
@@ -89,6 +89,8 @@ def _save_credentials(credentials):
     with open(CREDENTIALS_FILE, "w") as fd:
         os.chmod(fd.name, 0o600)
         json.dump(credentials, fd, sort_keys=True, indent=2)
+        _CREDENTIALS.clear()
+        _CREDENTIALS.update(credentials)
 
 
 def get_context_admin_user(context_id):  # type: (Union[int, str]) -> str